+++ tasks/modules/defaults.ts
@@ -5,6 +5,7 @@ import * as utils from './utils';
 const TypeScriptDefaults: IGruntTSOptions = {
     allowBool: false,
     allowImportModule: false,
+    allowSyntheticDefaultImports: null,
     amdloader: null,
     compile: true,
     declaration: false,
@@ -33,8 +34,12 @@ const TypeScriptDefaults: IGruntTSOptions = {
     emitGruntEvents: null,
     noEmitOnError: false,
     preserveConstEnums: false,
+    suppressExcessPropertyErrors: false,
     suppressImplicitAnyIndexErrors: false,
+    stripInternal: false,
     noEmit: false,
+    noLib: false,
+    emitBOM: false,
     inlineSources: false,
     inlineSourceMap: false,
     newLine: utils.eol,
@@ -43,6 +48,7 @@ const TypeScriptDefaults: IGruntTSOptions = {
     additionalFlags: '',
     templateCache: null,
     targetName: '',
+    locale: null,
     jsx: null,
     moduleResolution: null,
     experimentalAsyncFunctions: null,
@@ -66,3 +72,4 @@ function applyGruntTSDefaults(options: IGruntTSOptions) {
   options.emitGruntEvents = false;
   return options;
 }
+