+++ test/expected/htmlTemplate/out.js
@@ -1,3 +1,4 @@
+/* tslint:disable:max-line-length */
 var advanced;
 (function (advanced) {
     var $$template;
@@ -13,6 +14,8 @@ var advanced;
 })(advanced || (advanced = {}));
 /// <reference path="../reference.ts"/>
 var boo = advanced.$$template.name.tpl_html_module.advanced_$$template_name_tpl_html_variable;
+//grunt-start
 /// <reference path="src/advanced.$$template.name.tpl.html.ts" />
 /// <reference path="src/foo.ts" />
+//grunt-end 
 //# sourceMappingURL=out.js.map