+++ test/htmlSnakeModuleName/reference.ts
@@ -1,0 +1,4 @@
+//grunt-start
+/// <reference path="src/snake-module-name.tpl.html.ts" />
+/// <reference path="src/foo.ts" />
+//grunt-end