+++ test/htmlSnakeModuleName/src/foo.ts
@@ -1,0 +1,3 @@
+/// <reference path="../reference.ts"/>
+
+var boo = snakeModuleName.tpl_html_module.snakeModuleName_tpl_html_variable;