+++ test/test.js
@@ -65,6 +65,7 @@ exports.tests = {
     html2ts: function (test) {
         testDirectory(test, 'html');
         testDirectory(test, 'htmlTemplate');
+        testDirectory(test, 'htmlSnakeModuleName');
         test.done();
     },
     index: function (test) {